+++ tinyserver/src/client/mailclient.ts
@@ -1,4 +1,4 @@
-import { SERVER_HOSTNAME, SMTP_PASSWORD, SMTP_USER, SMTPClient } from 'tinyserver/deps.ts';
+import { ENV, SMTPClient } from 'tinyserver/deps.ts';
 
 const mailcli = new SMTPClient({
   debug: {
@@ -8,8 +8,8 @@ const mailcli = new SMTPClient({
     hostname: 'mailer',
     port: 1025,
     auth: {
-      username: SMTP_USER,
-      password: SMTP_PASSWORD,
+      username: ENV.SMTP_USER,
+      password: ENV.SMTP_PASSWORD,
     },
   },
 });
@@ -18,7 +18,9 @@ export const sendMail = async (
   params: { to: string; fromInfo?: string; subject: string; content: string; html: string },
 ) => {
   const to = params.to;
-  const from = `${params.fromInfo ?? 'noreply'}@${SERVER_HOSTNAME === 'localhost' ? 'example.com' : SERVER_HOSTNAME}`;
+  const from = `${params.fromInfo ?? 'noreply'}@${
+    ENV.SERVER_HOSTNAME === 'localhost' ? 'example.com' : ENV.SERVER_HOSTNAME
+  }`;
   await mailcli.send({
     from,
     to,
