+++ tinyserver/src/client/s3client.ts
@@ -1,16 +1,9 @@
-import {
-  BUCKET_ACCESS_KEY_ID,
-  BUCKET_ENDPOINT,
-  BUCKET_NAME,
-  BUCKET_REGION,
-  BUCKET_SECRET_KEY,
-  S3Bucket,
-} from 'tinyserver/deps.ts';
+import { ENV, S3Bucket } from 'tinyserver/deps.ts';
 
 export const bucket = new S3Bucket({
-  accessKeyID: BUCKET_ACCESS_KEY_ID,
-  secretKey: BUCKET_SECRET_KEY,
-  bucket: BUCKET_NAME,
-  region: BUCKET_REGION,
-  endpointURL: BUCKET_ENDPOINT,
+  accessKeyID: ENV.BUCKET_ACCESS_KEY_ID,
+  secretKey: ENV.BUCKET_SECRET_KEY,
+  bucket: ENV.BUCKET_NAME,
+  region: ENV.BUCKET_REGION,
+  endpointURL: ENV.BUCKET_ENDPOINT,
 });
