+++ tinyserver/src/model/ConfirmingEmailAddress.ts
@@ -1,6 +1,6 @@
+import { bcrypt, compareAsc, ENV } from 'tinyserver/deps.ts';
 import { type DBConfirmingEmailAddress, type DBUser, prisma } from 'tinyserver/src/client/dbclient.ts';
 import { sendMail } from 'tinyserver/src/client/mailclient.ts';
-import { bcrypt, compareAsc, SERVER_URI } from 'tinyserver/deps.ts';
 import { uniqueKey, uniqueSequentialKey } from 'tinyserver/src/utils/uniqueKey.ts';
 
 export class ConfirmingEmailAddress {
@@ -29,7 +29,7 @@ export const addEmailConfirmation = async (
   const token = uniqueKey();
   const hashedToken = bcrypt.hash(token);
 
-  const confirmURLBuilder = new URL(`${SERVER_URI}/confirm`);
+  const confirmURLBuilder = new URL(`${ENV.SERVER_URI}/confirm`);
   confirmURLBuilder.searchParams.append('token', `${id}:${token}`);
   confirmURLBuilder.searchParams.append('expired_at', expired_at.toJSON());
   const confirmURL = confirmURLBuilder.toString();
