+++ tinyserver/src/model/Files.ts
@@ -1,5 +1,5 @@
-import { type DBFiles, prisma } from 'tinyserver/src/client/dbclient.ts';
 import { v4, z } from 'tinyserver/deps.ts';
+import { type DBFiles, prisma } from 'tinyserver/src/client/dbclient.ts';
 import { bucket } from 'tinyserver/src/client/s3client.ts';
 import { User } from './Users.ts';
