+++ tinyserver/src/utils/fido2.ts
@@ -1,10 +1,10 @@
-import { Fido2Lib, SERVER_HOSTNAME, SERVER_URI, z } from '../../deps.ts';
+import { ENV, Fido2Lib, z } from '../../deps.ts';
 
 export const f2l = new Fido2Lib({
   timeout: 42,
-  rpId: SERVER_HOSTNAME,
+  rpId: ENV.SERVER_HOSTNAME,
   rpName: 'E2EEENCLOUD',
-  rpIcon: `${SERVER_URI}/logo.png`,
+  rpIcon: `${ENV.SERVER_URI}/logo.png`,
   challengeSize: 128,
   attestation: 'none',
   cryptoParams: [-7, -257],
